Move package-merge
