+++ src/lib/package-merge/index.js
